Split resource nav into separate file, grouped resource files
