+++ Source/Resource/ResourceObserver.swift
